refactor: :recycle: Move camera functionality into separate module
